# openAPI.yml
openapi: 3.0.0
info:
  title: cab booking api
  version: 1.0.0
servers:
  - url: 'http://localhost:3000'
paths:
  /users/sign-up:
    post:
      summary: Signup
      operationId: post-users-signup
      responses:
        '200':
          description: Sign up success
        '400':
          description: |-
            Invalid email
            Invalid phone number
            Name must contain at least 3 character
            Minimum password length is 10 characters and no spaces allowed.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                phone:
                  type: string
              required:
                - name
                - email
                - password
                - phone
            examples:
              invalid:
                value:
                  email: testgmail.com
                  password: '3949'
                  name: as
                  phone: sdf34
              valid:
                value:
                  email: test2@gmail.com
                  password: '1234567890'
                  name: Name
                  phone: '12345678'
      description: sign up new user
    parameters: []
  /users/login:
    post:
      summary: Login
      operationId: post-user-login
      responses:
        '200':
          description: Login success
          headers:
            Authorizartion:
              schema:
                type: string
              description: Authorization token
        '400':
          description: Password not provided
        '401':
          description: Invalid credentials
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
            examples:
              invalid:
                value:
                  email: test2@gmail.com
              valid:
                value:
                  email: test2@gmail.com
                  password: '1234567890'
      description: login user
    parameters: []
  /cabs/near-by:
    get:
      summary: Query near by cabs
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    cab:
                      type: object
                      properties:
                        status:
                          type: string
                        driver_id:
                          type: number
                        number_plate:
                          type: string
                        number_of_seats:
                          type: number
                        longitude:
                          type: string
                        latitude:
                          type: string
                        id:
                          type: number
                    distance:
                      type: number
        '400':
          description: Current location not set
      operationId: get-user-nearByCabs
      parameters:
        - schema:
            type: string
          in: header
          name: Authorization
        - schema:
            type: boolean
          in: query
          name: ignore_booked
        - schema:
            type: number
          in: query
          name: number_of_seats
      description: get near by cabs
    parameters: []
  /cabs/book:
    post:
      summary: Book Cab
      operationId: post-cab-book
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  bookingID:
                    type: string
              examples:
                example-1:
                  value:
                    message: Booking confirmed
                    bookingID: 5e93fc9de9395e22bc4dd7ab
        '400':
          description: No matching cabs
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cabID:
                  type: string
                pickupLocation:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                destination:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                numberOfPassengers:
                  type: number
            examples:
              valid:
                value:
                  cabID: 81
                  pickupLocation:
                    latitude: 40.7271
                    longitude: -74.0054
                  destination:
                    latitude: 40.7271
                    longitude: -72.0054
                  numberOfPassengers: 2
      description: book the specified cab
    parameters: []
  /users/current/bookings:
    get:
      summary: Query user bookings
      tags: []
      operationId: get-user-bookings
      parameters:
        - schema:
            type: string
          in: header
          name: Authorization
      description: get bookings
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    pickup_latitude:
                      type: string
                    pickup_longitude:
                      type: string
                    destination_latitude:
                      type: string
                    destination_longitude:
                      type: string
                    user_id:
                      type: number
                    fare:
                      type: string
                    number_of_passengers:
                      type: number
                    status:
                      type: string
                    commute_status:
                      type: string
                    cab_id:
                      type: string
                    created_at:
                      type: string
                    updated_at:
                      type: string
              examples:
                example:
                  value:
                    - id: 1
                      pickup_latitude: '40.7271'
                      pickup_longitude: '-74.0054'
                      destination_latitude: '40.7271'
                      destination_longitude: '-72.0054'
                      user_id: 2
                      fare: 337059
                      number_of_passengers: 2
                      status: CONFIRMED
                      commute_status: NOT_STARTED
                      cab_id: 81
                      created_at: '2020-05-07T15:58:55.000Z'
                      updated_at: null
    parameters: []
  /users/current/location:
    post:
      summary: Set current location
      operationId: post-user-setCurrentLocation
      responses:
        '200':
          description: OK
        '400':
          description: latitude and longitude must be provided and must be numeric.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
            examples:
              valid:
                value:
                  latitude: 40.7271
                  longitude: -74.0054
              invalid:
                value:
                  latitude: 40.7271
      description: set current location
    parameters: []
components:
  schemas: {}